pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pysocks>=1.7.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "setuptools>=80.9.0",
    "soupsieve>=2.7",
    "sqlalchemy>=2.0.42",
]
